requirements.txt
# Core dependencies
python-telegram-bot==20.7
aiohttp==3.9.3
SQLAlchemy==2.0.27
aiosqlite==0.19.0
python-dotenv==1.0.1
structlog==24.1.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Monitoring and Metrics
prometheus-client==0.19.0
psutil==5.9.8
fastapi==0.95.2
uvicorn==0.27.1

# Web Scraping and Analysis
selenium==4.18.1
fake-useragent==1.4.0
beautifulsoup4==4.12.3

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Logging
python-json-logger==2.0.7

# Captcha Solving
2captcha-python==1.2.0

# Added from the code block
pydantic==1.10.13
httpx==0.25.2
alembic==1.13.1
lxml==5.1.0
